Simple on off temperature ctrl is working

+++ lib/Display/ScreenStuff.h
@@ -7,11 +7,11 @@
   #include "Free_Fonts.h" // Include the header file attached to this sketch
 
   void Draw_screen_background();
-  void Draw_data(float , float , float , float , float , int , int , float , float , float );
-  void Init_Table(int TempTable[][TABLE_HEIGHT]);
-  void Add_Value_Table(int TempTable[][TABLE_HEIGHT],float , float , float );
+  void Draw_data(float , float , float , float ,  int , int , float ,float , float , float );
+  void Init_Table(uint16_t TempTable[][TABLE_HEIGHT]);
+  void Add_Value_Table(uint16_t TempTable[][TABLE_HEIGHT],float , float , float );
   void Display_UpTime(unsigned long );
   void Display_Refresh(unsigned long);
-  void Draw_Table(int TempTable[][TABLE_HEIGHT]);
+  void Draw_Table(uint16_t TempTable[][TABLE_HEIGHT]);
 
 #endif
